+++ src/find.js
@@ -3,7 +3,6 @@
 /*::
 import type {ImmutableTree} from './types.js';
 */
-import findBy from './findBy.js';
 
 export default function find/*:: <T, K = T, D = T> */(
   tree/*: ImmutableTree<T> */,
@@ -11,9 +10,16 @@ export default function find/*:: <T, K = T, D = T> */(
   cmp/*: (a: K, b: T) => number */,
   defaultValue/*: D */,
 )/*: T | D */ {
-  return findBy/*:: <T, D> */(
-    tree,
-    (treeValue) => cmp(key, treeValue),
-    defaultValue,
-  );
+  let cursor = tree;
+  while (cursor.size !== 0) {
+    const order = cmp(key, cursor.value);
+    if (order === 0) {
+      return cursor.value;
+    } else if (order < 0) {
+      cursor = cursor.left;
+    } else {
+      cursor = cursor.right;
+    }
+  }
+  return defaultValue;
 }