+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
   * Empty trees are now represented by `empty` (a tree of size 0) rather
     than `null`. Many function type signatures were updated to reflect this.
+  * Renamed `withComparator` to `withKeyComparator`.
+    The passed-in comparator now compares keys of type `K`. A second
+    argument is now required to retrieve `K` from `T`. This allows using
+    `K` in `update`, `find`, `findNext`, `findPrev`, and `indexOf`.
   * Added `validate` to check if a tree is valid for a given comparator.
   * Changed the type signature of `equals` to allow comparing trees of
     different types.